Rename a frontend component for consistency
