+++ frontend/src/components/rule-edit/tracking-rule/ArtifactsGroupOwnedParams.vue
